updated to v0.2-2

+++ src/DynamicModel.h
@@ -385,8 +385,10 @@ sampleBeta(const Size j,
                                next_b / next_sg2);
 
     /* Construct object of LogDenPar_type */
-    struct IntRegModel<Prior, Par>::LogDenPar data = {ldp_mu, ldp_sg2, this->N_,
-             ldp_X, ldp_dleY
+    struct IntRegModel<Prior, Par>::LogDenPar data = {
+	ldp_mu, ldp_sg2, 
+	static_cast<int> (this->N_),
+	ldp_X, ldp_dleY
     };
     double xsamp = 0.0;
 
@@ -491,8 +493,10 @@ sampleBeta(const Size j,
                                next_b / next_sg2);
 
     /* Construct object of LogDenPar_type */
-    struct IntRegModel<Prior, Par>::LogDenPar data = {ldp_mu, ldp_sg2, this->N_,
-             ldp_X, ldp_dleY
+    struct IntRegModel<Prior, Par>::LogDenPar data = {
+	ldp_mu, ldp_sg2, 
+	static_cast<int> (this->N_),
+	ldp_X, ldp_dleY
     };
     double xsamp = 0.0;
