+++ src/TimeIndepModel.h
@@ -87,8 +87,10 @@ sampleBeta(const ublas::vector<double>& lambda,
     }
 
     /* Construct object of LogDenPar_type */
-    struct IntRegModel<Prior, Par>::LogDenPar data = {ldp_mu, ldp_sg2, this->N_,
-             ldp_X, ldp_dleY
+    struct IntRegModel<Prior, Par>::LogDenPar data = {
+	ldp_mu, ldp_sg2, 
+	static_cast<int> (this->N_),
+	ldp_X, ldp_dleY
     };
     double xsamp = 0.0;
