+++ src/TimeVaryingModel.h
@@ -119,8 +119,10 @@ sampleBeta(const ublas::vector<double>& tLambda,
                                  next_b / next_sg2);
 
       /* Construct object of LogDenPar_type */
-      struct IntRegModel<Prior, Par>::LogDenPar data = {ldp_mu, ldp_sg2, this->N_,
-               ldp_X, ldp_dleY
+      struct IntRegModel<Prior, Par>::LogDenPar data = {
+	  ldp_mu, ldp_sg2, 
+	  static_cast<int> (this->N_),
+	  ldp_X, ldp_dleY
       };
       double xsamp = 0.0;
 
@@ -210,8 +212,10 @@ sampleBeta(const ublas::vector<double>& tLambda,
                                  next_b / next_sg2);
 
       /* Construct object of LogDenPar_type */
-      struct IntRegModel<Prior, Par>::LogDenPar data = {ldp_mu, ldp_sg2, this->N_,
-               ldp_X, ldp_dleY
+      struct IntRegModel<Prior, Par>::LogDenPar data = {
+	  ldp_mu, ldp_sg2, 
+	  static_cast<int> (this->N_),
+	  ldp_X, ldp_dleY
       };
       double xsamp = 0.0;
